+++ History.md
@@ -1,8 +1,9 @@
-unreleased
-==========
+1.0.0 / 2014-08-10
+==================
 
   * Accept valid `Vary` header string as `field`
   * Add `vary.append` for low-level string manipulation
+  * Move to `jshttp` orgainzation
 
 0.1.0 / 2014-06-05
 ==================